Move component DataView from 'components' to 'components/index/_shared'
