+++ components/index/CardsMonitoring/UntrackedRate/Chart.vue
@@ -111,13 +111,13 @@ import Vue from 'vue'
 import { TranslateResult } from 'vue-i18n'
 import { ThisTypedComponentOptionsWithRecordProps } from 'vue/types/options'
 
-import DataView from '@/components/DataView.vue'
 import DataViewDataSetPanel from '@/components/DataViewDataSetPanel.vue'
 import DataViewTable, {
   TableHeader,
   TableItem,
 } from '@/components/DataViewTable.vue'
 import ScrollableChart from '@/components/index/_shared/charts/_base/ScrollableChart.vue'
+import DataView from '@/components/index/_shared/DataView.vue'
 import {
   DisplayData,
   yAxesBgPlugin,
